api/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
azure-core==1.11.0
    # via azure-storage-blob
azure-storage-blob==12.7.1
    # via -r requirements.in
certifi==2020.12.5
    # via
    #   msrest
    #   requests
cffi==1.14.4
    # via cryptography
chardet==4.0.0
    # via requests
click==7.1.2
    # via
    #   typer
    #   uvicorn
cryptography==3.4.4
    # via azure-storage-blob
fastapi==0.63.0
    # via -r requirements.in
h11==0.12.0
    # via uvicorn
idna==2.10
    # via requests
isodate==0.6.0
    # via msrest
loguru==0.5.3
    # via -r requirements.in
msrest==0.6.21
    # via azure-storage-blob
numpy==1.20.1
    # via
    #   pandas
    #   patsy
    #   scipy
    #   statsmodels
oauthlib==3.1.0
    # via requests-oauthlib
pandas==1.2.2
    # via
    #   -r requirements.in
    #   statsmodels
patsy==0.5.1
    # via statsmodels
pyconfs[toml]==0.5.1
    # via -r requirements.in
pycparser==2.20
    # via cffi
pydantic==1.7.3
    # via fastapi
pyjwt==2.0.1
    # via -r requirements.in
pyplugs==0.3.2
    # via pyconfs
python-dateutil==2.8.1
    # via pandas
python-dotenv==0.15.0
    # via -r requirements.in
pytz==2021.1
    # via pandas
requests-oauthlib==1.3.0
    # via msrest
requests==2.25.1
    # via
    #   -r requirements.in
    #   azure-core
    #   msrest
    #   requests-oauthlib
scipy==1.6.0
    # via statsmodels
six==1.15.0
    # via
    #   azure-core
    #   isodate
    #   patsy
    #   python-dateutil
starlette==0.13.6
    # via fastapi
statsmodels==0.12.2
    # via -r requirements.in
toml==0.10.2
    # via pyconfs
typer==0.3.2
    # via -r requirements.in
urllib3==1.26.3
    # via requests
uvicorn==0.13.3
    # via -r requirements.in
